fix(docs-infra): optimize updateFrame by not using activeStyles
activeStyles had to exist and be incorporated into the updateFrame logic with the old Styles model where we didn't have "from" styles. They are obsolete now and actually increase the computational cost of updateFrame. Not only that, they are causing issues when you go backward in time (activeStyles supersede "to" styles irrespective whether they should or shouldn't; they shouldn't in the case where "to" styles are closer to currentTime compared to activeStyles). This is why, they are now being dropped. Respectively, curr/retreat styles are renamed to source styles to match the new updateFrame calculations changes.